Core: update params and results files (#504)
* Core: Fix Unchecked Type Assertion (#1305)

This commit fixes an unchecked type assertion when extracting results
into a slice pointer. The type assertion would previously fail when a
paged result was empty.

* Core: Fix issue with decoding slices of composed structs (#1964)

* Prevent Recursive BuildRequestBody

This commit prevents golangsdk from performing a recursive
BuildRequestBody when the field's JSON tag is "-".

* Fix required tag for queries and headers (#913)

This commit fixes how the required tag is set for query and header
parameters. Instead of using

x:"name,required"

as before, it now uses

x:"name" required:"true"

* update query tag format

* BuildRequestQuery to check slice of struct fields

This patch enhances BuildRequestQuery so that in case you have a
required struct field that is defined as a slice of structs, all the
structs in the slice will be checked. This makes it consistent with the
current behavior for struct fields.

* Core: Support for int64 headers (#1860)